refactor: move wrap_inputs_for_backward_compatibility to generic location
